requirements.txt
torch>=2.5.0
torchvision>=0.20.0
numpy>=1.26.0
pandas>=2.2.0
PyYAML>=6.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
seaborn>=0.13.0
tqdm>=4.66.0
Pillow>=10.0.0
thop>=0.1.1.post2209072238
